修改清理垃圾图片脚本

+++ inotify_imagemagic_rsync_log_cleanold/cleanrubbishpic.sh
@@ -26,6 +26,9 @@ getrubbishbydiff() {
 
 ##处理文件移动到备份区
 movetobackup() {
+    if [ ! -f $alreadypath/$curlock ]; then
+        return 0
+    fi
     for dofile in $(cat $alreadypath/$curlock); do
         if [ -f $actionpath/$dofile ]; then
             mv $actionpath/$dofile $backuppath/$dofile
@@ -35,7 +38,9 @@ movetobackup() {
 
 ##将处理完的文件锁加上后缀名 表示完成
 addexttolock() {
-    mv $alreadypath/$curlock $alreadypath/$curlock$curlockext
+    if [ -f $alreadypath/$curlock ]; then
+        mv $alreadypath/$curlock $alreadypath/$curlock$curlockext
+    fi
 }
 
 ##获取未操作过的
@@ -107,12 +112,17 @@ curlockext=.lock
 
 case "$1" in
     day)
+        if [ "" != "$2" ]; then
+            operatedate=$2
+            usingfile=$usingpath/$operatedate$usingext
+            curlock=tmp_lock_$operatedate
+        fi
         getoperatefileslist;
-        sleep 3;
+        sleep 2;
         getrubbishbydiff
         sleep 1;
         movetobackup
-        sleep 3;
+        sleep 2;
         addexttolock
         rm -f /tmp/$operatedate /tmp/2_$operatedate
         echo "done";
@@ -121,17 +131,17 @@ case "$1" in
         for dodate in $(ls $usingpath|cut -d '.' -f 1); do
             curdate=$(date "+%Y%m%d");  #建议放在这里 每次执行的时候都重新赋值一次当前日期  避免后半夜执行 跨过12点
             if [ "$dodate" = "$curdate" ]; then
-                continue;
+                continue 1;
             fi
             operatedate=$dodate;
             usingfile=$usingpath/$operatedate$usingext
             curlock=tmp_lock_$operatedate
             getoperatefileslist;
-            sleep 3;
+            sleep 2;
             getrubbishbydiff;
             sleep 1;
             movetobackup;
-            sleep 3;
+            sleep 2;
             addexttolock;
             rm -f /tmp/$operatedate
         done
@@ -162,16 +172,25 @@ case "$1" in
         getalreadybyday
         ;;
        *)
-        echo "
-        ***必须依托于usingfile目录中的列表文件***
-        Usage 使用方法 $0  
-        day 前一天的 
-        days 今天之前的全部日期 
-        mark 指定要删除的文件内容(第二个参数是比对文件名 :./todo.txt) 
-        didnot 查看比对文件中没有操作的 
-        checkday 查看某一天的操作记录(第二个参数是日期:20151212)
-        ";
-        exit 1;
+        tput clear
+        tput cup 3 15
+        tput setaf 3
+        echo "简介"
+
+        tput cup 6 15
+        echo "day. 前一天的  第二个参数是指定的日期(如果不指定默认取昨天)"
+        tput cup 7 15
+        echo "mark. 指定要删除的文件内容(第二个参数是比对文件名 :./todo.txt) "
+        tput cup 8 15
+        echo "didnot. 查看比对文件中没有操作的 "
+        tput cup 9 15
+        echo "checkday 查看某一天的操作记录(第二个参数是日期:20151212)"
+        tput sgr0
+        tput bold
+        tput cup 5 17
+        echo "***必须依托于usingfile目录中的列表文件***"
+        tput sgr0
+        tput rc
         ;;
 esac
